# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "myProperty API",
        "description": "developed by Jonathan",
        "contact": {
            "email": "jonathanx56@hotmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:80",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/v1/search": {
            "get": {
                "tags": [
                    "real-states-search"
                ],
                "summary": "Get Search Real-state for external use",
                "operationId": "Search Real-state",
                "parameters": [
                    {
                        "name": "conditions",
                        "in": "query",
                        "description": "-   Follow the format below attribute:caracter:value",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "-   Enter the name of the fields separated by commas",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/v1/search/{real_state_id}": {
            "get": {
                "tags": [
                    "real-states-search"
                ],
                "summary": "Show real-states with andress and photo for external use",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "real_state_id",
                        "in": "path",
                        "description": "-   Enter the real_state_id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "List Users",
                "operationId": "list",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create User",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "-   Enter the name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "-   Enter the email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "-   Enter the password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "-   Enter the password_confirmation",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "profile[phone]",
                        "in": "query",
                        "description": "-   Enter the property profile",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Show User",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "-   Enter the id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete User",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "-   Enter the id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        }
    },
...
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}